Add `install_ads` scripts to automate conda env creation and dependency installation (#821)
* Copy `environment.yml` deps into `requirements.txt`

* `requirements.txt`: Unpin skimage to avoid slow wheel build

* `requirements.txt`: Limit `numpy<2` to avoid torch error
See: https://stackoverflow.com/a/78881990

* `cp install_sct install_ads`

* `install_ads`: Swap SCT steps for ADS steps

* `install_ads`: Swap "SCT"/"spinalcordtoolbox" strs for "ADS"

* `install_ads`: Standardize on `ads_conda` for conda dir

For some reason, historically SCT has always used `python/` as the name for the conda
installation dir. I've always found this to be a bit confusing, so now is as good a time
as ever to change this for ADS.

* `.gitignore`: Add newly-generated files

* `install_ads`: Remove unnecessary "data" directory steps

* `install_ads`: Remove unnecessary binary install step

* `install_ads`: Remove unnecessary model downloading step

* `install_ads`: Remove unnecessary old SCT installation check

* `install_ads`: Remove SCT-specific GPU torch installer option

SCT installs CPU torch by default, with GPU torch being opt-in. However, since ADS uses
GPU torch by default, this option has no use.

* `cp install_sct.bat install_ads.bat`

* `install_ads.bat`: Swap SCT steps for ADS steps

* `install_ads.bat`: Swap "SCT"/"spinalcordtoolbox" strs for "ADS"

* `install_ads.bat`: Standardize on `ads_conda` for conda dir

For some reason, historically SCT has always used `python/` as the name for the conda
installation dir. I've always found this to be a bit confusing, so now is as good a time
as ever to change this for ADS.

* `run_tests.yaml`: Update to use `install_ads` script

* `test_compute_morphometrics.py`: Add call to `np.abs`

For some reason, some of the results are negatives of the other? TODO: Discuss why this might be on PR.

* `__init__.py`: Make sure repo dir is on `PATH`

For some reason, on Windows, the repo directory doesn't get added to the PATH. This
means that `config.py` cannot be found, and `pytest` throws errors.

* Run CI when merging to the nnunet branch

* Fix PR wildcard for CI runs

* Update test_compute_morphometrics.py

* Update download_tests.py

* Reupload morphometrics file

* Update install_ads for plugins install

* Update install_ads.bat  for plugins install

* Update setup.cfg

* Add alias for napari and activating ADS, and expanded user instructions at the end of install

* First draft for adding `ads_activate`/`ads_napari` entry points

* Revert "First draft for adding `ads_activate`/`ads_napari` entry points"

This reverts commit 8bb8cdab0662b25741a57957ac65f4d811d63e33.

* Use `bin/` scripts to launch `activate` and `napari`

* Update documentation with installation instructions

* Fix syntax

* Add 99% of Joshua's suggested changes

* Add windows CLI path instructions

* Update documentation.rst

---------

Co-authored-by: Mathieu Boudreau, PhD <emb6150@gmail.com>